requirements.txt
Django==5.2.6
Pillow==11.1.0
django-widget-tweaks==1.5.0
django-redis==6.0.0
redis==6.4.0
deep-translator==1.11.4
google-generativeai==0.8.4
beautifulsoup4==4.14.2
requests==2.32.3